Crate aquariwm_server

AquariWM is an ICCCM- and EWMH-compliant window manager for the X Window System. It focuses on providing a modular approach to window management, allowing different functionality to be ‘mixed and matched’ according to a user’s wishes. It supports this with a large effort to provide clear and extensive documentation and guides for users, contributors and clientmakers alike, and by maintaining a friendly and welcoming community that encourages participation.

Modularity

AquariWM’s core feature is its modularity. To understand how it achieves this, you must first understand the way in which the X Window System works.

X Window System

The X window system is centered around a client-server model, as you may be used to using for the internet. Each graphical application that runs on your computer under X is known as an X client. When your computer is running an X server, these clients can communicate with the server, sending various messages and requests, and receiving various notifications and replies back.

A window manager fills a role somewhere in between these two sides: it requests that certain requests made by clients are redirected to itself, rather than to the X server. A window manager can send and receive messages to and from X clients, and it can decide which messages to allow, modify, or deny, before sending its own messages to the X server.

AquariWM clients

In very much a similar way to how X clients communicate with the X server (or any window manager that may be intercepting these messages), AquariWM allows its modularity by defining a protocol for X clients to communicate directly with AquariWM. AquariWM’s functionality can be split across multiple ‘AquariWM clients’, allowing an extreme level of control over how the window manager operates.

What does this allow AquariWM to do differently?

With a typical window manager, there is a limit to how much any particular functionality can be customized without making direct changes to the window manager itself. In contrast, features can be added or replaced in AquariWM without affecting any other part of the window manager. Changing the way part of AquariWM works is as simple as launching a different component.

Modules

aquariwm 🔒

The central module of aquariwm-server. Responsible for overseeing AquariWM’s state and operation.

extensions

This module provides an assortment of utility traits to ease interaction with xcb.

features 🔒

The features module contains various other modules for functionality of AquariWM.

util

Contains utilities for sending X requests, especially for queries or initialization on windows.

Functions

main 🔒

The main entrypoint for AquariWM; main is responsible for initalization/setup.