Enum aquariwm_server::features::window_manipulation::WindowManipulation

pub enum WindowManipulation {
    Moving {
        window: Window,
        parent: Window,
        orig_coords: (i16, i16),
        cursor_pos: (i16, i16),
    },
    Resizing {
        window: Window,
        orig_size: (u16, u16),
        cursor_pos: (i16, i16),
    },
}

Represents the state of an ongoing window manipulation.

Contains the window() being manipulated, the cursor_pos() relative to the root window when the manipulation commenced, and the original position or size of the window() to calculate the offset and to allow the window() to be returned to its original state with WindowManipulation::cancel().

Only one window can be manipulated at one time, and only one mode of window manipulation may operate on that window at one time. If the window is being moved, it cannot also be resized in the same window manipulation.

Variants

Moving

Fields

window: Window

The window this window manipulation applies to.

parent: Window

The current parent window of the window being manipulated.

The Moving window manipulation keeps track of the window’s parent because window coordinates are relative to the window’s parent window, not the root window. Since the window manipulation stores the original position (orig_coords) of the window, the coordinates must be translated from the parent window’s coordinates to the window’s new parent if the window happens to be reparented during manipulation.

orig_coords: (i16, i16)

The original position of the window, relative to its parent.

If the window being manipulated is reparented during the course of its manipulation, these coordinates should be translated from the old parent to the new one.

cursor_pos: (i16, i16)

The starting position of the cursor when the window manipulation begins.

The cursor_pos is relative to the window’s root window.

This is used to calculate how far the cursor has moved during window manipulation, and thus the offset that should be applied to the window’s position.

The window()’s position is being manipulated; the user is moving the window().

Resizing

Fields

window: Window

The window this window manipulation applies to.

orig_size: (u16, u16)

The original dimensions of the window.

cursor_pos: (i16, i16)

The starting position of the cursor when the window manipulation begins.

The cursor_pos is relative to the window’s root window.

This is used to calculate how far the cursor has moved during window manipulation, and thus the offset that should be applied to the window’s dimensions.

The window()’s dimensions are being manipulated; the user is resizing the window().

Implementations

impl WindowManipulation

pub fn moving(
conn: &Connection,
window: Window,
cursor_pos: (i16, i16)
) -> Result<Self>

Creates a WindowManipulation::Moving window manipulation.

Blocking: waits for one GetGeometry request, one QueryTree request.

Flushes the connection.

pub fn resizing(
conn: &Connection,
window: Window,
cursor_pos: (i16, i16)
) -> Result<Self>

Creates a WindowManipulation::Resizing window manipulation.

Blocking: waits for one GetGeometry request.

Flushes the connection.

pub const fn is_moving(self) -> bool

Returns whether this WindowManipulation is Moving.

This is the inverse of WindowManipulation::is_resizing(), and is simply another shorthand for:

matches!(self, Self::Moving { .. })

See also: matches!

pub const fn is_resizing(self) -> bool

Returns whether this WindowManipulation is Resizing.

This is the inverse of WindowManipulation::is_moving(), and is simply another shorthand for:

matches!(self, Self::Resizing { .. })

See also: matches!

pub fn cancel(self, conn: &Connection) -> Result<()>

Returns the window() being manipulated to its original position or size.

This does not end the window manipulation. Make sure you don’t continue to manipulate the window if you want this to have an effect.

Flushes the connection.

pub const fn window(self) -> Window

The window this window manipulation applies to.

pub const fn cursor_pos(self) -> (i16, i16)

The starting position of the cursor when the window manipulation begins.

See also: cursor_x(), cursor_y()

pub const fn cursor_x(self) -> i16

The x coordinate of the cursor when the window manipulation begins.

See also: cursor_pos(), cursor_y()

pub const fn cursor_y(self) -> i16

The y coordinate of the cursor when the window manipulation begins.

See also: cursor_pos(), cursor_y()

pub const fn diff(self, cursor_pos: (i16, i16)) -> (i16, i16)

The difference between the given cursor position and the starting cursor_pos().

See also: diff_x(x: i16), diff_y(y: i16)

pub const fn diff_x(self, x: i16) -> i16

The difference between the given x coordinate and the starting cursor_x().

See also: diff(cursor_pos: (i16, i16)), diff_y(y: i16)

pub const fn diff_y(self, y: i16) -> i16

The difference between the given y coordinate and the starting cursor_y().

See also: diff(cursor_pos: (i16, i16)), diff_x(x: i16)

pub fn apply(self, conn: &Connection, cursor_pos: (i16, i16)) -> Result<()>

Applies the window manipulation to the given window.