Module aquariwm_server::util

Contains utilities for sending X requests, especially for queries or initialization on windows.

Functions

get_geometry

Sends a GetGeometry request for the given window and returns its reply.

grab_manip_buttons

Grabs button motion and releases on the given window.

init_grabs

Grabs button presses on the given window when the Super key is held.

init_window

Initializes the given window by requesting to receive certain events on it.

set_dimensions

Sends a ConfigureWindow request to change the dimensions of the given window.

set_position

Sends a ConfigureWindow request to change the coordinates of the given window.